pyproject.toml
[project]
name = "grimoire-engine"
version = "0.1.0"
description = "Dynamic RPG world simulation backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "aiosqlite>=0.20.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
grimoire-cli = "grimoire.cli.harness:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["grimoire"]
